Hydrodynamics:
Started development of a new hydrodynamics implementation.
The new implementation will use the same buoyancy forces.
But for the hydrodynamics (drag etc) it will use Fossen textbook model.
I am using the approach taken by the VRX platform.
With this, I am rewriting the buoyancy part of the algorithm.
Theres now a WaterInteraction library which helps structure the scripts.
See Buoyancy.cs for how this new library is used.
I will integrate this into the previous implementation to make it nicer.
For the new dynamics, there will be one script for buoyancy.
And another for the damping, coriolis etc.
Since these dont rely on submersion information.